# backend/apiSpec.yaml
openapi: 3.1.0
info:
  title: Home Automation
  description: |-
    This is a documentation for the home automation api to be used with the hardware
  contact:
    email: ahmadujuwon@gmail.com
  license:
    name: GNU GPLv3
    url: https://choosealicense.com/licenses/gpl-3.0/
  version: 0.0.1
...
paths:   
...
  /groups:
    get:
      tags:
        - Group
      summary: Gets all groups and nodes of authenticated user
      responses:
        "200":
          description: Groups fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      groups:
                        type: array
                        items:
                          $ref: "#/components/schemas/Group"
    post:
      tags:
        - Group
      summary: Creates a new group
      requestBody:
        content:
          application/json: 
            schema:
              required:
                - name
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/Group/properties/name'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json: 
              schema:
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      group:
                        $ref: '#/components/schemas/Group'
        '400':
          description: Group creation unsuccessful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request body validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidBodyErrorResponse'
...
  /admin/nodes:
      get:
        tags:
          - Admin
        summary: Get all User's nodes
        parameters:
          - in: query
            name: user_id
            # required: true
            schema:
                description: User id to fetch nodes for
                type: string
          - in: query
            name: not_activated
            # required: true
            schema:
                description: Fetch nodes that haven't been activated
                type: boolean
        responses:
          '200':
            description: Nodes fetched successfully
            content:
              application/json: 
                schema:
                  properties:
                    success:
                      type: boolean
                      enum: [true]
                    message:
                      type: string
                    data:
                      properties:
                        nodes:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                $ref: "#/components/schemas/Node/properties/id"
                              name:
                                $ref: "#/components/schemas/Node/properties/name"
          '403':
            description: User not admin
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
      post:
        tags:
          - Admin
        summary: Create a new node
        requestBody:
          content:
            application/json: 
              schema:
                required:
                  - type
                type: object
                properties:
                  type:
                    $ref: '#/components/schemas/Node/properties/type'
        responses:
            '201':
              description: Node created successfully
              content:
                application/json: 
                  schema:
                    properties:
                      success:
                        type: boolean
                        enum: [true]
                      message:
                        type: string
                      data:
                        properties:
                          node:
                            type: object
                            properties:
                              id:
                                $ref: "#/components/schemas/Node/properties/id"
                              type:
                                $ref: '#/components/schemas/Node/properties/type'
            '400':
              description: Node creation unsuccessful
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/ErrorResponse'
            '422':
              description: Request body validation failed
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/InvalidBodyErrorResponse'
            '403':
              description: User not admin
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users
      responses:
        '200':
          description: Users fetched successfully
          content:
            application/json: 
              schema:
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  data:
                    properties:
                      users:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              $ref: "#/components/schemas/User/properties/id"
                            first_name:
                              $ref: "#/components/schemas/User/properties/first_name"
                            last_name:
                              $ref: "#/components/schemas/User/properties/last_name"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User not admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        nodes:
          type: array
          items:
              $ref: '#/components/schemas/GroupNodeResponse'
    InvalidBodyErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
              field:
                type: string
    ErrorResponse:
      type: object
      properties:
          success:
              type: boolean
              enum: [false]
          message:
              type: string
...
    GroupNodeResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Node/properties/id'
        name:
          $ref: '#/components/schemas/Node/properties/name'
        is_online:
          $ref: '#/components/schemas/Node/properties/is_online'
        state:
          $ref: '#/components/schemas/Node/properties/state'
        last_state_change:
          $ref: '#/components/schemas/Node/properties/last_state_change'
        state_duration:
          $ref: '#/components/schemas/Node/properties/state_duration'